Refactor : modify folder name
