+++ src/App.js
@@ -2,9 +2,9 @@ import "./App.css";
 
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
-import Login from "./Login";
-import SignUp from "./SignUp";
-import Home from "./Home";
+import Login from "./page/Login";
+import SignUp from "./page/SignUp";
+import Home from "./page/Home";
 
 function App() {
     return (
