+++ src/page/Home.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+function Home(){
+
+    return(
+        <div>
+            홈
+        </div>
+    )
+}
+
+export default Home;
